@sbinet 2021-06-24
FYI, this has been integrated into https://github.com/go-pdf/fpdf.